requirements.txt
eve
flask
requests
pyjwt